% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.R
\name{ncbi_pubmed}
\alias{ncbi_pubmed}
\title{Search PubMed articles}
\usage{
ncbi_pubmed(query, retmax = 20, api_key = NULL)
}
\arguments{
\item{query}{Search query (e.g. "CRISPR AND 2024[pdat]")}

\item{retmax}{Max results (default 20)}

\item{api_key}{Optional API key}
}
\value{
tibble: uid, name (title), description (journal)
}
\description{
Search PubMed articles
}
