% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.R
\name{ncbi_summary}
\alias{ncbi_summary}
\title{Get summaries for NCBI IDs}
\usage{
ncbi_summary(db, ids, api_key = NULL)
}
\arguments{
\item{db}{Database name}

\item{ids}{Character vector of NCBI IDs (from ncbi_search)}

\item{api_key}{Optional API key}
}
\value{
tibble: uid, name, description (columns vary by database)
}
\description{
Get summaries for NCBI IDs
}
